// src/app/interfaces/clients.interface.ts
export interface Clients {
    id: number;
    name: string;
    lastName: string;
    email: string;
    age: string;
    genre: string;
    lat: number; 
    long: number;
    distance: string;
    city: string;
    country: string;
    postalCode: string;
    phone: string;
    income: number;
    status: string;
}